build: drop gtkdocize
The inclusion of gtk-doc.make in docs/Makefile.am is breaking autoreconf
when we're used as a submodule; we don't want to force users to depend
on GTKDoc, especially not when we won't even use it by default.

This commit adds the Makefile gtkdocize provides to the repo. We also
add an empty file in m4/ so it's ready for aclocal when we're first
cloned.