gtu/test-case: soften assert message
Given that users are meant to signal GTU about test failure via
gtu_assert, the message we print is a bit over-the-top and could
possibly break some test harnesses. So in the default case of non-fatal
asserts, we instead print a more informative message that should be
handled properly by the GLib test logging infrastructure.