src/log: fix deadlock when running dummy suite
Attempting to run a test suite with no child tests results in the
logging of an empty test plan which itself results in the logging of an
empty test plan due to an atexit handler. Since the log function is
guarded by g_once_init_enter(), the test program will log the test plan
and then hang.

This commit makes sure to call g_once_init_leave() before exiting and
adds a simple reproducer to test-progs.