src/log/glib: drop fatal domain testing
The primary GTU logging domain is exposed to glib-handlers.c in order
for that code to make decisions about how to handle logged fatal
messages, but the log handling policy was kind of confusing and also
broken as of 6fba8ff7.

Since it doesn't currently do anything, it's removed to simplify the
code.