docs: hack around gtkdoc-check
The docbook makes liberal use of includes to make life a bit easier, but
gtkdoc-check insists some of the content files aren't used because it
assumes all used files are included in the main file. It does this check
by iterating over the lines of the main file (gtu-docs.xml, in this
instance) and checking if any lines have the relative path of a content
file as a substring.

Since we know these files are used, we just name them in comments at the
end of the file. Hopefully I remember to update this whenever the set of
content files change...