docs: document log handling policy
Explicitly outline what GTU does when a message is logged via GLib's
logging system and why.

We also add G-DEBUG:CAPS as the ID for the relevant varlist entry to
allow links. GLib's docbook doesn't define it so it should be fine for
now, but this might cause issues in the future.